Key signature picking
You can now select the key signature you want to generate notes from
with the drop down menu. For the console, the third argument is now the
key signature you would like represented as an integer between 0 and 50
inclusive and the fourth argument is whether or not you want sound
output.